+++ addon/utils/shared/value.js
@@ -7,14 +7,14 @@ export function receiveValue(selectBox) {
   updateValue(selectBox, selectBox.args.value);
 }
 
-export async function updateValue(selectBox, value) {
-  await resolveValue(selectBox, value, processValue);
-  updatedValue(selectBox);
+export function updateValue(selectBox, value) {
+  return resolveValue(selectBox, value, processValue).then(() =>
+    updatedValue(selectBox)
+  );
 }
 
-export async function selectValue(selectBox, value) {
-  await updateValue(selectBox, value);
-  selectedValue(selectBox);
+export function selectValue(selectBox, value) {
+  return updateValue(selectBox, value).then(() => selectedValue(selectBox));
 }
 
 function processValue(selectBox, value) {