+++ tests/integration/components/select-box-test.js
@@ -204,8 +204,6 @@ module('select-box', function (hooks) {
 
     this.set('myValue', value2);
 
-    await settled();
-
     assert.deepEqual(
       apis[1].value,
       ['bar'],